requirements.txt
aiohttp==3.7.4.post0
async-timeout==3.0.1
attrs==22.2.0
dateutils==0.6.12
discord.py==2.2.2
rfc3987==1.3.8
mysql-connector-python